Update Notification. Notification is changed to single component and Other thing are also fixed.
