+++ src/core/Notification/Notification.baseStyles.ts
@@ -1,13 +1,29 @@
 import { css } from 'styled-components';
-import { SuomifiTheme } from '../../theme';
-import { font } from '../../theme/reset';
-import { baseNotificationBaseStyles } from '../BaseNotification/BaseNotification.baseStyles';
+import { SuomifiTheme } from '../theme';
+import { font, element } from '../theme/reset';
 
 export const baseStyles = (theme: SuomifiTheme) => css`
-  ${baseNotificationBaseStyles(theme)}
   &.fi-notification {
-    padding-bottom: ${theme.spacing.l};
+    ${element(theme)}
+    ${font(theme)('bodyTextSmall')}
+    padding-bottom: 17px;
+    width: 100%;
+    box-shadow: ${theme.shadows.notificationBoxShadow};
+    border-radius: 4px;
 
+    & .fi-notification_text-content-wrapper {
+      margin-top: 22px;
+      padding-left: 20px;
+      & .fi-notification_content {
+        vertical-align: middle;
+        ${font(theme)('bodyTextSmall')}
+      }
+    }
+    & .fi-notification_label {
+      ${font(theme)('bodySemiBold')}
+      margin-top: -1px;
+      margin-bottom: ${theme.spacing.xxs};
+    }
     & .fi-notification_action-element-wrapper {
       margin: 10px 20px 5px 93px;
       & .fi-button {
@@ -27,12 +43,14 @@ export const baseStyles = (theme: SuomifiTheme) => css`
       display: flex;
       flex-wrap: nowrap;
       box-sizing: border-box;
-      margin-top: ${theme.spacing.s};
+      margin-top: 14px;
       margin-right: -5px;
       margin-bottom: ${theme.spacing.insetM};
     }
     & .fi-notification_style-wrapper {
       padding: 0 43px 10px 43px;
+      display: flex;
+      align-items: flex-start;
     }
     & .fi-notification_close-button {
       ${font(theme)('bodyTextSmall')}
@@ -41,6 +59,7 @@ export const baseStyles = (theme: SuomifiTheme) => css`
       padding: 7px 8px;
       border: 1px solid transparent;
       border-radius: ${theme.radius.basic};
+      text-transform: uppercase;
 
       &:focus-visible {
         outline: 0;
@@ -65,12 +84,33 @@ export const baseStyles = (theme: SuomifiTheme) => css`
         transform: translateY(1px);
       }
     }
+    /* Status variant styles */
+    &--neutral {
+      border-top: 4px solid ${theme.colors.accentSecondary};
+      & .fi-notification_icon-wrapper .fi-icon .fi-icon-base-fill {
+        fill: ${theme.colors.accentSecondary};
+      }
+    }
+    &--error {
+      border-top: 4px solid ${theme.colors.alertBase};
+      & .fi-notification_icon-wrapper .fi-icon .fi-icon-base-fill {
+        fill: ${theme.colors.alertBase};
+      }
+    }
     /** Small screen variant styles */
     &.fi-notification--small-screen {
       padding-bottom: 20px;
       & .fi-notification_text-content-wrapper {
+        margin-top: 16px;
+        padding-top: 1px;
         padding-left: 12px;
         padding-right: 12px;
+        display: flex;
+        flex: 3 1 auto;
+        flex-direction: column;
+      }
+      & .fi-notification_label {
+        margin-top: -2px;
       }
       & .fi-notification_close-button-wrapper {
         justify-content: flex-end;
