--- src/core/Notification/index.ts
@@ -1,5 +1,0 @@
-export { Notification, NotificationProps } from './Notification/Notification';
-export {
-  InlineNotification,
-  InlineNotificationProps,
-} from './InlineNotification/InlineNotification';
