+++ src/index.tsx
@@ -6,7 +6,10 @@ export {
   BreadcrumbLinkProps,
 } from './core/Breadcrumb';
 export { Alert, AlertProps, InlineAlert, InlineAlertProps } from './core/Alert';
-export { Notification, NotificationProps } from './core/Notification';
+export {
+  Notification,
+  NotificationProps,
+} from './core/Notification/Notification';
 export { Block, BlockProps } from './core/Block/Block';
 export { Button, ButtonProps } from './core/Button/Button';
 export { Dropdown, DropdownProps } from './core/Dropdown/';
